# docs/static/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for AltMount — a WebDAV/NFS server backed by NZB/Usenet, with queue management, file health monitoring, and media streaming.",
        "title": "AltMount API",
        "contact": {
            "name": "AltMount GitHub Issues",
            "url": "https://github.com/javi11/altmount/issues"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/javi11/altmount/blob/master/LICENSE"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/import/nzbdav": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Starts an async import from an NzbDav SQLite database file (by path or upload)",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "NZBDav"
                ],
                "summary": "Import NzbDav database",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Root folder path for imported files",
                        "name": "rootFolder",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Server-side path to SQLite database file",
                        "name": "dbPath",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "SQLite database file upload (if dbPath not provided)",
                        "name": "file",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Cancels any in-progress NzbDav import operation",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "NZBDav"
                ],
                "summary": "Cancel NzbDav import",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/import/nzbdav/reset": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Resets the NzbDav import status to allow starting a new import",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "NZBDav"
                ],
                "summary": "Reset NzbDav import status",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/import/nzbdav/status": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the current status of the NzbDav import operation",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "NZBDav"
                ],
                "summary": "Get NzbDav import status",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "api.APIMeta": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "api.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/api.APIError"
                },
                "meta": {
                    "$ref": "#/definitions/api.APIMeta"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "API key for programmatic access (available from GET /api/config)",
            "type": "apiKey",
            "name": "apikey",
            "in": "query"
        },
        "BearerAuth": {
            "description": "JWT token obtained from POST /api/auth/login. Format: \"Bearer \u003ctoken\u003e\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}